Add max_depth parameter
When I was using the profiler with more complex examples I realized that
drilling down too much is not very useful and the output is hard to
read. Being able to limit it gives you the option to run it shallow at
first and then decide how much yoy want to drill down.